+++ app/resources/views/pages/dominion/magic.blade.php
@@ -144,7 +144,7 @@
                                                     $canCast = $spellCalculator->canCast($selectedDominion, $spell);
                                                 @endphp
                                                 <div class="col-xs-6 col-sm-3 col-md-6 col-lg-3 text-center">
-                                                    <div class="form-group">
+                                                    <div class="form-group war-non-chaos">
                                                         <button type="submit"
                                                                 name="spell"
                                                                 value="{{ $spell->key }}"
@@ -167,6 +167,31 @@
                                                             @endif
                                                         </small>
                                                     </div>
+                                                    @if ($isBlackGuard)
+                                                        <div class="form-group war-chaos" style="display: none;">
+                                                            <button type="submit"
+                                                                    name="spell"
+                                                                    value="{{ $spell->key }}"
+                                                                    class="btn btn-primary btn-block war-spell disabled"
+                                                                    {{ $selectedDominion->isLocked() || $selectedDominion->round->hasOffensiveActionsDisabled() || !$canCast || (now()->diffInDays($selectedDominion->round->start_date) < 3) ? 'disabled' : null }}>
+                                                                {{ $spellHelper->getChaosSpellName($spell) }}
+                                                            </button>
+                                                            <p style="margin: 5px 0;">{{ $spellHelper->getChaosSpellDescription($spell) }}</p>
+                                                            <small>
+                                                                @if ($canCast)
+                                                                    Mana cost: <span class="text-success">{{ number_format($spellCalculator->getManaCost($selectedDominion, $spell)) }}</span><br/>
+                                                                @else
+                                                                    Mana cost: <span class="text-danger">{{ number_format($spellCalculator->getManaCost($selectedDominion, $spell)) }}</span><br/>
+                                                                @endif
+                                                                @if ($spell->duration)
+                                                                    Lasts {{ $spell->duration }} hours<br/>
+                                                                @endif
+                                                                @if (!empty($spell->races))
+                                                                    Racial<br/>
+                                                                @endif
+                                                            </small>
+                                                        </div>
+                                                    @endif
                                                 </div>
                                             @endforeach
                                         </div>
@@ -237,7 +262,7 @@
                                                     $buttonStyle = ($isActive ? 'btn-success' : 'btn-primary');
                                                 @endphp
                                                 <div class="form-group">
-                                                    <button type="submit" name="spell" value="{{ $spell->key }}" class="btn {{ $buttonStyle }} btn-block" {{ $selectedDominion->isLocked() || !$canCast || $cooldownHours || ($selectedDominion->protection_ticks_remaining && $spell->hasPerk('invalid_protection')) ? 'disabled' : null }}>
+                                                    <button type="submit" name="spell" value="{{ $spell->key }}" class="btn {{ $buttonStyle }} btn-block" {{ $selectedDominion->isLocked() || !$canCast || $cooldownHours || ($selectedDominion->protection_ticks_remaining && $spell->hasPerk('invalid_protection')) || (!$isBlackGuard && in_array('chaos-league', $spell->races)) ? 'disabled' : null }}>
                                                         {{ $spell->name }}
                                                     </button>
                                                     <p style="margin: 5px 0;">{{ $spellHelper->getSpellDescription($spell) }}</p>
@@ -357,6 +382,13 @@
                 var revengeStatus = $(this).find(":selected").data('revenge');
                 var guardStatus = $(this).find(":selected").data('guard');
                 var friendlyStatus = $(this).find(":selected").data('friendly');
+                if (guardStatus == 1) {
+                    $('.war-non-chaos').hide();
+                    $('.war-chaos').show();
+                } else {
+                    $('.war-chaos').hide();
+                    $('.war-non-chaos').show();
+                }
                 if (!friendlyStatus && (warStatus == 1 || revengeStatus == 1 || guardStatus == 1)) {
                     $('.war-spell').removeClass('disabled');
                 } else {