+++ src/Helpers/SpellHelper.php
@@ -33,13 +33,13 @@ class SpellHelper
             ->active()
             ->get()
             ->map(function ($spell) {
-                $spell->racial = ($spell->races !== []);
+                $spell->racial = ($spell->races !== [] && $spell->races !== ['chaos-league']);
                 return $spell;
             });
 
         if ($race !== null) {
             $spells = $spells->filter(function ($spell) use ($race) {
-                if (empty($spell->races) || in_array($race->key, $spell->races)) {
+                if (!$spell->racial || in_array($race->key, $spell->races)) {
                     return true;
                 }
                 return false;
@@ -217,7 +217,8 @@ class SpellHelper
             'wizard_power_defense' => '%+g%% defensive wizard power',
             'wonder_damage' => 'Deals damage to wonders',
             'explore_cost_wizard_mastery' => 'Exploring platinum cost reduced by 1%% per %d Wizard Mastery (max 10%%)',
-            'spell_refund' => 'Failed spells refund %d%% of their mana cost',
+            'spell_refund' => 'Failed chaos spells refund %d%% of their strength and mana costs',
+            'invalid_royal_guard' => 'Cannot be cast while in the Royal Guard',
             'apply_rejuvenation' => 'Applies Rejuvenation upon expiration',
             'immune_burning' => 'Immune to Burning',
             'immune_lightning_storm' => 'Immune to Lightning Storm',
@@ -246,6 +247,30 @@ class SpellHelper
         }
 
         return implode($separator, $perkStrings);
+    }
+
+    public function getChaosSpellName(Spell $spell): string
+    {
+        switch ($spell->key) {
+            case 'fireball':
+                return 'Chaos Fireball';
+            case 'lightning_bolt':
+                return 'Chaos Lightning';
+            case 'disband_spies':
+                return 'Chaos Disband';
+        }
+    }
+
+    public function getChaosSpellDescription(Spell $spell): string
+    {
+        switch ($spell->key) {
+            case 'fireball':
+                return 'Kills 6% unprotected peasants';
+            case 'lightning_bolt':
+                return 'Temporarily destroys 0.3% science, keep, forges, walls';
+            case 'disband_spies':
+                return 'Turns 2% of spies into random resources for yourself';
+        }
     }
 
     public function getSpellRaces(Spell $spell, string $separator = ', '): string